+++ R/log_app_usage.R
@@ -106,7 +106,7 @@ line_usage_log <- function(...) {
 line_pkg_log <- function(fields) {
   nest_packages <- c(
     "rtables", "hermes", "teal.modules.hermes",
-    "tern", "teal", "teal.devel", "teal.modules.general", "teal.modules.clinical",
+    "tern", "teal", "teal.modules.general", "teal.modules.clinical",
     "osprey", "teal.osprey", "goshawk", "teal.goshawk"
   )
   pkg_desc <- utils::sessionInfo()$otherPkgs