+++ R/module_tabs_with_filters.R
@@ -37,7 +37,9 @@
 #'     )
 #'   },
 #'   server = function(input, output, session) {
-#'     active_module <- srv_tabs_with_filters(id = "dummy", datasets = datasets, modules = mods, filter = NULL)
+#'     active_module <- srv_tabs_with_filters(
+#'       id = "dummy", datasets = datasets, modules = mods, filter = NULL
+#'     )
 #'     output$info <- renderText({
 #'       paste0("The currently active tab name is ", active_module()$label)
 #'     })
@@ -65,8 +67,12 @@
 #'     )
 #'   },
 #'   server = function(input, output, session) {
-#'     active_module1 <- teal:::srv_tabs_with_filters(id = "app1", datasets = datasets1, modules = mods, filter = NULL)
-#'     active_module2 <- teal:::srv_tabs_with_filters(id = "app2", datasets = datasets2, modules = mods, filter = NULL)
+#'     active_module1 <- teal:::srv_tabs_with_filters(
+#'       id = "app1", datasets = datasets1, modules = mods, filter = NULL
+#'     )
+#'     active_module2 <- teal:::srv_tabs_with_filters(
+#'       id = "app2", datasets = datasets2, modules = mods, filter = NULL
+#'     )
 #'     output$info <- renderText({
 #'       paste0(
 #'         "The currently active tab names are: ",
@@ -149,7 +155,7 @@ srv_tabs_with_filters <- function(id, datasets, modules, filter) {
     )
 
     datasets$srv_filter_panel(id = "filter_panel", active_datanames = active_datanames)
-    set_filter_state(datasets = datasets, filter = filter)
+    teal.slice::set_filter_state(datasets = datasets, filter = filter)
     showNotification("Data loaded - App fully started up")
 
     logger::log_trace(