+++ R/teal.logger-exports.R
@@ -1,3 +1,0 @@
-#' @importFrom teal.logger register_logger
-#' @export
-teal.logger::register_logger