+++ NEWS.md
@@ -1,4 +1,4 @@
-# teal 0.10.1.9018
+# teal 0.10.1.9021
 
 * Added `metadata` field to `TealDataset` to store a named list of `metadata` items. It is available for module developers through `FilteredData$get_metadata("<<dataname>>")` and can be pulled or added directly to datasets derived from `TealDatasetConnectors`.
 
@@ -8,6 +8,7 @@
 * Add counts to filtering categorical variables bar charts in the filtering panel in cases where they were missing.
 
 ### Breaking changes
+* `default_filter` has been deprecated - use `list()` instead for a default filter.
 * Due to deprecation of `root_modules` any `label` argument to `modules` must be explicitly named. For example `modules("lab", mod1, mod2)` should be replaced with `modules(label = "lab", mod1, mod2)`.
 * Minor changes to the interface of `TealDataset`, for example some active fields should be replaced by explicit "get" calls.